src/requirements.txt
click < 6.7
pyyaml < 3.13
jinja2~=2.9
docker-py < 1.11
dockerpty < 0.5
templated-yaml >= 1.1.1
